Demo/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
pydantic_settings>=2.0.0
python-dotenv>=1.0.0
motor>=3.3.2
pymongo>=4.6.1
langchain>=0.3.21
langchain-anthropic>=0.3.10
langchain-core>=0.3.49 